requirements.txt
ffmpeg
moviepy
requests
discord-ext-bot
websocket-client